+++ lib/elm-test.js
@@ -25,7 +25,6 @@ var compile = require("node-elm-compiler").compile,
   fs = require("fs-extra"),
   crypto = require("crypto"),
   glob = require("glob"),
-  builder = require("xmlbuilder"),
   path = require("path"),
   util = require("util"),
   _ = require("lodash"),
@@ -201,7 +200,7 @@ function runTests(testFile) {
       }
       prepareCompiledJsFile(initialSeed, report, dest);
 
-      Supervisor.run(dest);
+      Supervisor.run(dest, args.watch);
     }
   });
 }
@@ -224,7 +223,7 @@ function prepareCompiledJsFile(initialSeed, report, dest) {
     content,
     "return module.exports;",
     "})({});",
-    "var initialSeed = " + initialSeed + ";",
+    "var initialSeed = " + String(initialSeed) + ";",
     'var report = "' + report + '";',
     after
   ].join("\n");
