+++ lib/supervisor.js
@@ -2,9 +2,10 @@
 
 var os = require("os"),
   chalk = require("chalk"),
+  builder = require("xmlbuilder"),
   child_process = require("child_process");
 
-function run(dest) {
+function run(dest /*:string*/, watch /*:boolean*/) {
   var cpus = os.cpus() || new Array(1);
 
   var nextTestToRun = -1;
@@ -12,7 +13,7 @@ function run(dest) {
   var finishedWorkers = 0;
   var closedWorkers = 0;
   var results = [];
-  var testsToRun = 4; // TODO record this number from BEGIN
+  var testsToRun = 4; // TODO: record this number from BEGIN
 
   function flushResults() {
     // Only print any results if we're ready - that is, nextResultToPrint
@@ -39,7 +40,7 @@ function run(dest) {
     nextTestToRun++;
 
     // Immediately run the next test.
-    worker.send("" + testToRun);
+    worker.send({ message: String(testToRun) });
   }
 
   var workers = cpus.map(function(throwaway, index) {
@@ -64,7 +65,7 @@ function run(dest) {
           // If all the workers have finished, print the summmary.
           if (finishedWorkers === workers.length) {
             // TODO send enough summary data that one worker can do this.
-            worker.send("SUMMARY");
+            worker.send({ message: "SUMMARY" });
           }
           break;
         case "SUMMARY":
@@ -77,6 +78,10 @@ function run(dest) {
           workers.forEach(function(worker) {
             worker.kill();
           });
+
+          if (!watch) {
+            process.exit(data.exitCode);
+          }
           break;
         case "BEGIN":
           // TODO record all the relevant values, and display the "hey we're
@@ -127,10 +132,10 @@ function run(dest) {
       // The BEGIN message requests metadata about the test run, e.g.
       // how many tests will be run, whether they should auto-fail because
       // of skip/on/y/todo, etc.
-      worker.send("BEGIN");
+      worker.send({ message: "BEGIN" });
     } else {
       // Send the index of the test to run.
-      worker.send("" + testToRun);
+      worker.send({ message: String(testToRun) });
     }
   });
 }
@@ -143,10 +148,6 @@ function printResult(data) {
       console.log(builder.create(data.message).end());
     } else {
       console.log(JSON.stringify(data.message));
-    }
-
-    if (!args.watch) {
-      process.exit(data.exitCode);
     }
   } else if (data.type === "STARTED" || data.type === "TEST_COMPLETED") {
     if (data.format === "CHALK") {
@@ -181,6 +182,22 @@ function chalkify(messages) {
       }
     })
     .join("");
+}
+
+// Fix Windows Unicode problems. Credit to https://github.com/sindresorhus/figures for the Windows compat idea!
+var windowsSubstitutions = [
+  [/[↓✗►]/g, ">"],
+  [/╵│╷╹┃╻/g, "|"],
+  [/═/g, "="],
+  ,
+  [/▔/g, "-"],
+  [/✔/g, "√"]
+];
+
+function windowsify(str) {
+  return windowsSubstitutions.reduce(function(result, sub) {
+    return result.replace(sub[0], sub[1]);
+  }, str);
 }
 
 function finish() {
